Corrected case of Xquery > XQuery
Great project! Looking forward to taking it for a spin!

(For citations on the case of XQuery, see https://www.w3.org/TR/xquery-3/ and https://en.wikipedia.org/wiki/XQuery. In my style book, it's "XQuery" or, in upper case-averse contexts, "xquery," but if upper case is allowed, it's never "Xquery" or "xQuery.")